wrote a testing environment runner for testing rules implemented simple tests for NamespaceContainsRule.cs and RequireBracesInBlockRule.cs
