+++ CodeAnalyzerTool/RoslynPlugin.Test/Rules/RequireBracesInBlockTest.cs
@@ -1,0 +1,45 @@
+﻿using Microsoft.CodeAnalysis;
+using RoslynPlugin.rules;
+
+namespace RoslynPlugin.Test;
+
+public class RequireBracesInBlockTest
+{
+    [Fact]
+    public async void RequiredBracesInBlockRule_ShouldReportRuleViolation_WhenIfStatementWithoutBracesExist()
+    {
+        var code = @"
+class C {
+    private void M() {
+        if (true) Console.WriteLine(""true is indeed true"") 
+    }
+}
+";
+        var rule = new RequireBracesInBlockRule();
+        rule.Options = new Dictionary<string, string>();
+        rule.Severity = DiagnosticSeverity.Warning;
+        
+        var results = await RuleTestRunner.CompileStringWithRule(code, rule);
+        results.Should().Contain(rv => rv.Rule.RuleName == RuleNames.REQUIRE_BRACES_IN_BLOCK_RULE);
+    }
+    
+    [Fact]
+    public async void RequiredBracesInBlockRule_ShouldNotReportRuleViolation_WhenIfStatementWithoutBracesDoesNotExist()
+    {
+        var code = @"
+class C {
+    private void M() {
+        if (true) {
+            Console.WriteLine(""true is indeed true"")
+        }
+    }
+}
+";
+        var rule = new RequireBracesInBlockRule();
+        rule.Options = new Dictionary<string, string>();
+        rule.Severity = DiagnosticSeverity.Warning;
+        
+        var results = await RuleTestRunner.CompileStringWithRule(code, rule);
+        results.Should().BeEmpty();
+    }
+}
