--- CodeAnalyzerTool/RoslynPlugin.Test/Verifier.cs
@@ -1,15 +1,0 @@
-﻿namespace RoslynPlugin.Test;
-
-public class Verifier
-{
-    // public static async Task VerifyAnalyzerAsync(string source, params DiagnosticResult[] expected)
-    // {
-    //     var test = new Test
-    //     {
-    //         TestCode = source,
-    //     };
-    //
-    //     test.ExpectedDiagnostics.AddRange(expected);
-    //     await test.RunAsync(CancellationToken.None);
-    // }
-}
