+++ CodeAnalyzerTool/RoslynPlugin/Rules/NamespaceContainsRule.cs
@@ -11,7 +11,7 @@ namespace RoslynPlugin.rules;
 [DiagnosticAnalyzer(LanguageNames.CSharp)]
 public class NamespaceContainsRule : RoslynRule
 {
-    public sealed override string RuleName => "NamespaceContains";
+    public sealed override string RuleName => RuleNames.NAMESPACE_CONTAINS_RULE;
     public sealed override DiagnosticSeverity Severity { get; set; }
     public sealed override Dictionary<string, string> Options { get; set; }
     public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; }
